feat(dashboard): prefer address names over raw addresses throughout UI
Extract the SenderCell address-display pattern into a reusable AddressLink
component and apply it consistently — pool list address column, pool detail
header (pool ID + token0/1), and breadcrumb loading fallback. Known addresses
and token symbols are shown as named links to the block explorer, with the full
0x address in the hover tooltip.

Made-with: Cursor
